src/fastapi/requirements.txt
boto3==1.26.133
fastapi==0.95.1
uvicorn==0.20.0
python-dotenv==0.21.0
SQLAlchemy==1.4.45
numpy==1.24.3
pandas==2.0.1
mlflow==2.3.1